+++ EssentialFeed/EssentialFeediOS/Feed UI/Composers/FeedLoaderPresentationAdapter.swift
@@ -1,0 +1,28 @@
+//
+//  Copyright © 2019 Essential Developer. All rights reserved.
+//
+
+import EssentialFeed
+
+final class FeedLoaderPresentationAdapter: FeedViewControllerDelegate {
+	private let feedLoader: FeedLoader
+	var presenter: FeedPresenter?
+	
+	init(feedLoader: FeedLoader) {
+		self.feedLoader = feedLoader
+	}
+	
+	func didRequestFeedRefresh() {
+		presenter?.didStartLoadingFeed()
+		
+		feedLoader.load { [weak self] result in
+			switch result {
+			case let .success(feed):
+				self?.presenter?.didFinishLoadingFeed(with: feed)
+				
+			case let .failure(error):
+				self?.presenter?.didFinishLoadingFeed(with: error)
+			}
+		}
+	}
+}
